setUp nav bar
