+++ src/components/views/hotel/aboutSection.js
